EngineNotBootingUpError fixed completely :: Initialization of any UGenericConfirmCancel class is only allowed in C++

+++ Source/TacticGame/Gameplay/Units/AdditionalComponents/UnitStatsComponent.h
@@ -78,7 +78,7 @@ public:
 	}
 
 	UFUNCTION(BlueprintCallable)
-	void AddXP(float XPValue)
+	void AddXP(int XPValue)
 	{
 		LevelProgression.AddXP(XPValue);
 	}
